# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pages/{slug}:
    get:
      operationId: getSupportPage
      tags: [pages]
      summary: Get a support page by slug
      description: Returns page details and all slots. For PIN-protected pages, requires the PIN to be passed as a query parameter.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: pin
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Support page found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportPageWithSlots"
        "401":
          description: PIN required or incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinRequiredError"
        "404":
          description: Page not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
...
components:
  schemas:
...
    SlotType:
      type: string
      enum:
        - meal
        - school_pickup
        - child_care
        - errand
        - dog_walking
        - shopping
        - visit
        - other
    SlotResponse:
      type: object
      properties:
        id:
          type: string
        pageId:
          type: string
        slotType:
          $ref: "#/components/schemas/SlotType"
        customLabel:
          type: string
          nullable: true
        slotDate:
          type: string
          nullable: true
          description: >-
            Null means the task has no fixed date — a flexible offer, claimed
            whenever suits. The date is set when a helper claims it.
        slotTime:
          type: string
          nullable: true
        liftWaitMode:
          $ref: "#/components/schemas/LiftWaitMode"
        notes:
          type: string
          nullable: true
        dietaryNotes:
          type: string
          nullable: true
          description: >-
            Meal slots only (bug #006): allergies / dietary preferences a helper
            needs before cooking. Null on every other slot type.
        headcount:
          type: integer
          nullable: true
          description: >-
            Meal slots only (bug #006): how many people the meal should feed.
            Null on every other slot type.
        isClaimed:
          type: boolean
        claimedByName:
          type: string
          nullable: true
        claimedNote:
          type: string
          nullable: true
        createdAt:
          type: string
        calendarUrl:
          type: string
          nullable: true
          description: >-
            webcal:// subscribe link to this claim's calendar feed. Populated
            ONLY on the response to POST /slots/{slotId}/claim, and only for a
            task with a date (an undated "whenever suits" offer isn't an
            appointment, so it's null). Absent/null everywhere else — the slot
            listing in GET /pages never sets it. Mirrors the trusted-invite
            claim response's calendarUrl field so both claim paths surface the
            same "Add to your calendar" link.
      required:
        - id
        - pageId
        - slotType
        - isClaimed
        - createdAt
    SupportPageWithSlots:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        recipientName:
          type: string
        situationDescription:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, pending_approval, active, closed]
        privacy:
          type: string
          enum: [open, pin_protected]
        goodToKnow:
          type: string
          nullable: true
          description: >-
            An optional free-text note from the recipient, shown to every
            helper. Null when they didn't leave one.
        helpingCount:
          type: integer
          description: >-
            Ambient presence — the number of distinct people helping (deduped by
            claimed contact), for the warm "N people helping" line. Carries the
            warmth without naming anyone, so it is safe to show to everyone.
        slots:
          type: array
          items:
            $ref: "#/components/schemas/SlotResponse"
      required:
        - id
        - slug
        - recipientName
        - status
        - privacy
        - helpingCount
        - slots
...
    PinRequiredError:
      type: object
      properties:
        error:
          type: string
        pinRequired:
          type: boolean
      required:
        - error
        - pinRequired
...
    NotFoundError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LiftWaitMode:
      type: string
      nullable: true
      description: >-
        Bug #033 — for a LIFT, whether the helper drops off, waits and brings
        them home, or collects. "Drop off" is a twenty-minute favour; "wait" can
        be half a day, and nothing used to say which.

        NULL IS MEANINGFUL AND IS THE COMMON CASE: it means "not a lift, or
        nobody has said yet", and every surface renders nothing at all for it. A
        dated "pick up a prescription" errand is null and must look untouched.

        The presence of this field is also what marks a task as a lift — there is
        no 'lift' slot type. See migration 0011.
      enum:
        - drop_off
        - wait
        - pick_up
...